+++ CHANGES.rst
@@ -9,3 +9,5 @@ Change log
 - Fetch all projects, not just the first 100.
 
 - Improve formatting to have always a precision of 2 for floats.
+
+- No longer ignore entries with negative time.